chore(db): fase 2 migracion postgres - tests contra postgres real

+++ backend/src/main/java/org/gipsybuho/recetasfamiliares/ratings/RecipeRatingEntity.java
@@ -37,7 +37,7 @@ public class RecipeRatingEntity {
     @JoinColumn(name = "user_id", nullable = false, columnDefinition = "varchar(36)")
     private UserEntity user;
 
-    @Column(nullable = false, columnDefinition = "smallint")
+    @Column(nullable = false, columnDefinition = "integer")
     private int stars;
 
     @Column(length = 500)
